requirements.txt
streamlit
pandas
openai
requests
matplotlib
scikit-learn
langchain
python-dotenv
httpx
plotly
